+++ corpus/labeled/hn_safe_read.js
@@ -1,0 +1,4 @@
+function h(configPath) {
+  return fs.readFileSync(configPath, "utf8");
+}
+module.exports = h;