+++ corpus/labeled/hn_snprintf.c
@@ -1,0 +1,4 @@
+#include <stdio.h>
+void fmt(char *out, size_t n, const char *s) {
+  snprintf(out, n, "%s", s);
+}