+++ corpus/labeled/tp_massassign.js
@@ -1,0 +1,4 @@
+function h(req) {
+  return new User(req.body);
+}
+module.exports = h;