+++ corpus/labeled/tp_pathtrav.js
@@ -1,0 +1,4 @@
+function h(req) {
+  return fs.readFileSync(req.params.file, "utf8");
+}
+module.exports = h;