+++ corpus/labeled/tp_ssrf.js
@@ -1,0 +1,4 @@
+async function h(req) {
+  return fetch(req.query.url);
+}
+module.exports = h;