+++ corpus/labeled/tp_weakcipher.js
@@ -1,0 +1,5 @@
+const crypto = require("crypto");
+function enc(key) {
+  return crypto.createCipher("aes192", key);
+}
+module.exports = enc;